requirement.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-filter==21.1
django-six==1.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
importlib-metadata==4.11.3
Markdown==3.3.6
PyJWT==2.3.0
pytz==2021.3
screen==1.0.1
sqlparse==0.4.2
xlwt==1.3.0
zipp==3.7.0
